Fix bug in c++ cube for size > 2^32 Added getKey function for cube in Scala Add new test case for Cube Add flag to indicate whether domain is increasing or decreasing

+++ srccpp/ds/Cube.h
@@ -20,7 +20,7 @@ struct BigArray {
     }
 
     double &operator[](long long n) {
-        int a = (int) n / P;
+        int a = n / P;
         int b = n % P;
         return data[a][b];
     }
